+++ backend/src/main/java/io/ssafy/p/j14c103/homerun/api/service/pass/response/PassWidgetResponse.java
@@ -2,14 +2,14 @@ package io.ssafy.p.j14c103.homerun.api.service.pass.response;
 
 public class PassWidgetResponse {
 
-    private final int todaySaved;
-    private final int weeklySaved;
-    private final int weeklyGoal;
+    private final long todaySaved;
+    private final long weeklySaved;
+    private final long weeklyGoal;
     private final int progressRate;
-    private final int remaining;
+    private final long remaining;
 
-    private PassWidgetResponse(final int todaySaved, final int weeklySaved,
-                               final int weeklyGoal, final int progressRate, final int remaining) {
+    private PassWidgetResponse(final long todaySaved, final long weeklySaved,
+                               final long weeklyGoal, final int progressRate, final long remaining) {
         this.todaySaved = todaySaved;
         this.weeklySaved = weeklySaved;
         this.weeklyGoal = weeklyGoal;
@@ -17,17 +17,17 @@ public class PassWidgetResponse {
         this.remaining = remaining;
     }
 
-    public static PassWidgetResponse of(final int todaySaved, final int weeklySaved, final int weeklyGoal) {
+    public static PassWidgetResponse of(final long todaySaved, final long weeklySaved, final long weeklyGoal) {
         final int rate = weeklyGoal > 0
                 ? (int) Math.min((double) weeklySaved / weeklyGoal * 100.0, 100.0)
                 : 0;
-        final int remaining = Math.max(weeklyGoal - weeklySaved, 0);
+        final long remaining = Math.max(weeklyGoal - weeklySaved, 0L);
         return new PassWidgetResponse(todaySaved, weeklySaved, weeklyGoal, rate, remaining);
     }
 
-    public int getTodaySaved() { return todaySaved; }
-    public int getWeeklySaved() { return weeklySaved; }
-    public int getWeeklyGoal() { return weeklyGoal; }
+    public long getTodaySaved() { return todaySaved; }
+    public long getWeeklySaved() { return weeklySaved; }
+    public long getWeeklyGoal() { return weeklyGoal; }
     public int getProgressRate() { return progressRate; }
-    public int getRemaining() { return remaining; }
+    public long getRemaining() { return remaining; }
 }
